feat: 구조 분해 및 비회원 추가 처리
비회원 추가 처리를 바텀 시트 페이지에서 분리했습니다

저장 후 검색 필드에 문자가 남아 있는 문제가 있어 초기화 시켜주었습니다
